+++ src/app/widgets.cpp
@@ -15,5 +15,6 @@
 #include "widgets/sparkline_widget.cpp"
 #include "widgets/table_widget.cpp"
 #include "widgets/rocker_widget.cpp"
+#include "widgets/numericrocker_widget.cpp"
 
 #endif // HAS_DISPLAY